requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
cryptography==36.0.1
fastapi==0.73.0
greenlet==1.1.2
h11==0.9.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
mysql-connector-python==8.0.28
packaging==21.3
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
